tempo

+++ Assets/scripts/MusicParser/Metronome.cs
@@ -3,9 +3,9 @@ public class Metronome
     internal Metronome()
     {
         BeatUnit = string.Empty;
-        PerMinute = 0;
+        PerMinute = 0.0f;
     }
 
     public string BeatUnit { get; internal set; }
-    public int PerMinute { get; internal set; }
+    public float PerMinute { get; internal set; }
 }
